+++ netbox_aws_vpc_plugin/tables.py
@@ -10,12 +10,13 @@ class AWSVPCTable(NetBoxTable):
     vpc_cidr = tables.Column(linkify=True)
     owner_account = tables.Column(linkify=True)
     region = tables.Column(linkify=True)
+    status = tables.Column()
     # TODO: Count of subnets
 
     class Meta(NetBoxTable.Meta):
         model = AWSVPC
-        fields = ("pk", "id", "vpc_id", "name", "arn", "vpc_cidr", "owner_account", "region", "actions")
-        default_columns = ("vpc_id", "name", "vpc_cidr", "owner_account", "region")
+        fields = ("pk", "id", "vpc_id", "name", "arn", "vpc_cidr", "owner_account", "region", "status", "actions")
+        default_columns = ("vpc_id", "name", "vpc_cidr", "owner_account", "region", "status")
 
 
 class AWSSubnetTable(NetBoxTable):
@@ -25,19 +26,33 @@ class AWSSubnetTable(NetBoxTable):
     subnet_cidr = tables.Column(linkify=True)
     owner_account = tables.Column(linkify=True)
     region = tables.Column(linkify=True)
+    status = tables.Column()
 
     class Meta(NetBoxTable.Meta):
         model = AWSSubnet
-        fields = ("pk", "id", "subnet_id", "vpc", "name", "arn", "subnet_cidr", "owner_account", "region", "actions")
-        default_columns = ("subnet_id", "name", "subnet_cidr", "vpc", "owner_account", "region")
+        fields = (
+            "pk",
+            "id",
+            "subnet_id",
+            "vpc",
+            "name",
+            "arn",
+            "subnet_cidr",
+            "owner_account",
+            "region",
+            "status",
+            "actions",
+        )
+        default_columns = ("subnet_id", "name", "subnet_cidr", "vpc", "owner_account", "region", "status")
 
 
 class AWSAccountTable(NetBoxTable):
     account_id = tables.Column(linkify=True)
     name = tables.Column()
     tenant = tables.Column(linkify=True)
+    status = tables.Column()
 
     class Meta(NetBoxTable.Meta):
         model = AWSAccount
-        fields = ("pk", "id", "account_id", "name", "arn", "tenant", "actions")
-        default_columns = ("account_id", "name", "tenant")
+        fields = ("pk", "id", "account_id", "name", "arn", "tenant", "status", "actions")
+        default_columns = ("account_id", "name", "tenant", "status")