+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.6 (pending)
+
+* Add `status` to the custom models
+* Add a table of the VPCs to the account view
+
 ## 0.0.5 (2025-04-20)
 
 * Add `region` to AWS Subnet and AWS VPC